Add typescript as a peerDependency
It appears that typescript needs to be installed, even in projects that
do not have files with .ts or .tsx extensions, or ESLint will complain
that it cannot be found.

If this becomes too annoying in the future, we could consider releasing
a version of this package that does not require typescript to be
installed.